git-svn-id: file:///vegas/users/staff/womo1998/.svnrepos/my-repos@63 a0bf5412-5ba6-11e5-acee-19d0f835ea7d

+++ trunk/PRCE.py
@@ -190,9 +190,9 @@ class PRCE(object):
         
         # Compute the modal solutions for all model orders
         
-        modal_frequencies = np.zeros((max_model_order, num_ref_channels*max_model_order/2))
-        modal_damping = np.zeros((max_model_order, num_ref_channels*max_model_order/2))
-        mode_shapes = np.ones((num_analised_channels, num_ref_channels*max_model_order/2, max_model_order),dtype=complex)
+        modal_frequencies = np.zeros((max_model_order, int(num_ref_channels*max_model_order/2)))
+        modal_damping = np.zeros((max_model_order, int(num_ref_channels*max_model_order/2)))
+        mode_shapes = np.ones((num_analised_channels, int(num_ref_channels*max_model_order/2), max_model_order),dtype=complex)
         
         #print("size of modal_frequencies = ", np.shape(modal_frequencies))
